requirements.txt
pyarrow>=10.0.1
Flask>=2.1.0
requests>=2.23.0
psycopg2-binary>=2.9.3
pyodbc>=4.0.39
snowflake-connector-python>=3.0.3
boto3>=1.26.145
gunicorn>=20.1.0